@@ -4,7 +4,7 @@ dotenv.config();
 
 const dbPool = mysql2.createPool({
   host: process.env.DB_HOST,
-  port: process.env.DB_PORT,
+  // port: process.env.DB_PORT,
   user: process.env.DB_USER,
   password: process.env.DB_PW,
   database: 'market-kurly-clone',
